+++ dflib/src/main/java/com/nhl/dflib/LongSeries.java
@@ -2,6 +2,8 @@ package com.nhl.dflib;
 
 import com.nhl.dflib.series.LongArraySeries;
 
+import java.util.Random;
+
 /**
  * A Series optimized to store and access primitive long values without <code>java.lang.Long</code> wrapper. Can also
  * pose as "Series&lt;Long>", although this is not the most efficient way of using it.
@@ -57,4 +59,16 @@ public interface LongSeries extends Series<Long> {
      * @return a LongSeries that contains non-repeating values from this Series.
      */
     LongSeries uniqueLong();
+
+    /**
+     * @since 0.7
+     */
+    @Override
+    LongSeries sample(int size);
+
+    /**
+     * @since 0.7
+     */
+    @Override
+    LongSeries sample(int size, Random random);
 }