+++ dflib/src/main/java/com/nhl/dflib/series/LongBaseSeries.java
@@ -11,6 +11,7 @@ import com.nhl.dflib.ValueMapper;
 import com.nhl.dflib.ValuePredicate;
 import com.nhl.dflib.concat.SeriesConcat;
 import com.nhl.dflib.groupby.SeriesGrouper;
+import com.nhl.dflib.sample.Sampler;
 import com.nhl.dflib.series.builder.BooleanAccumulator;
 import com.nhl.dflib.series.builder.IntAccumulator;
 import com.nhl.dflib.series.builder.LongAccumulator;
@@ -20,6 +21,7 @@ import com.nhl.dflib.series.builder.UniqueLongAccumulator;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Objects;
+import java.util.Random;
 
 /**
  * @since 0.6
@@ -125,6 +127,22 @@ public abstract class LongBaseSeries implements LongSeries {
         copyTo(sorted, 0, 0, size);
         Arrays.sort(sorted, comparator);
         return new ArraySeries<>(sorted);
+    }
+
+    private LongSeries selectAsLongSeries(IntSeries positions) {
+
+        int h = positions.size();
+
+        long[] data = new long[h];
+
+        for (int i = 0; i < h; i++) {
+            // unlike SelectSeries, we do not expect negative ints in the index.
+            // So if it happens, let it fall through to "getLong()" and fail there
+            int index = positions.getInt(i);
+            data[i] = getLong(index);
+        }
+
+        return new LongArraySeries(data);
     }
 
     private Series<Long> selectAsObjectSeries(IntSeries positions) {
@@ -439,6 +457,22 @@ public abstract class LongBaseSeries implements LongSeries {
     @Override
     public SeriesGroupBy<Long> group(ValueMapper<Long, ?> by) {
         return new SeriesGrouper<>(by).group(this);
+    }
+
+    /**
+     * @since 0.7
+     */
+    @Override
+    public LongSeries sample(int size) {
+        return selectAsLongSeries(Sampler.sampleIndex(size, size()));
+    }
+
+    /**
+     * @since 0.7
+     */
+    @Override
+    public LongSeries sample(int size, Random random) {
+        return selectAsLongSeries(Sampler.sampleIndex(size, size(), random));
     }
 
     @Override