+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #69 TableSaver: apply quotes to schema/catalog/table name separately
 * #70 Remove API deprecated in 0.6 
+* #71 Series, DataFrame "sample"
 
 ## Release 0.6